Refactor SliderButton styles for SliderBanner component. Add Home page. Refactor "App.tsx".
